refactor: split inputs.tsx into per-datatype modules under inputs/ (#31 part 2)
Splits the 629-line inputs.tsx into 13 small files under
src/components/inputs/. Pure refactor — every existing test passes
unchanged (258 / 258), and the public surface (defaultTypeInputs,
JsonFallbackInput, FhirTypeInput / FhirInputProps / InputContext /
TypeInputs) is preserved at the same import paths via the new
inputs/index.ts barrel.

Layout matches the structure called out in #31:

  components/inputs/
    index.ts           — assembles defaultTypeInputs + re-exports types
    types.ts           — shared types + form-field CSS classes
    primitives.tsx     — Text, Markdown, Boolean, Number, Date,
                         DateTime, Time, Uri, Code (binding-aware)
    HumanName.tsx, Address.tsx, ContactPoint.tsx, Identifier.tsx,
    Reference.tsx, Period.tsx, Quantity.tsx, Coding.tsx,
    CodeableConcept.tsx, JsonFallback.tsx

Internal import paths updated from "./inputs.js" to "./inputs/index.js"
in three places: components/index.ts, components/inputs.test.tsx,
components/ResourceEditor.tsx.

Adds .changeset/inputs-split.md (patch) — no behaviour change.

Closes #31 (combined with the format extraction in d13706d).
