+++ packages/react-fhir/src/components/index.ts
@@ -1,7 +1,7 @@
 export * from "./Narrative.js";
 export * from "./renderers.js";
 export * from "./ResourceView.js";
-export * from "./inputs.js";
+export * from "./inputs/index.js";
 export * from "./ResourceEditor.js";
 export * from "./ResourceSearch.js";
 export * from "./ReferencePicker.js";
