+++ packages/react-fhir/src/components/inputs.test.tsx
@@ -7,7 +7,7 @@ import { setupServer } from "msw/node";
 import { afterAll, afterEach, beforeAll, describe, expect, it, vi } from "vitest";
 import { FetchFhirClient } from "../client/FetchFhirClient.js";
 import { FhirClientProvider } from "../hooks/FhirClientProvider.js";
-import { defaultTypeInputs } from "./inputs.js";
+import { defaultTypeInputs } from "./inputs/index.js";
 
 const BASE = "https://fhir.example.test/fhir";
 const server = setupServer();
