+++ packages/react-fhir/src/components/inputs/Identifier.tsx
@@ -1,0 +1,27 @@
+import type { Identifier } from "fhir/r4";
+import { baseField, subLabel, type FhirTypeInput } from "./types.js";
+
+export const IdentifierInput: FhirTypeInput<Identifier> = ({ value, onChange }) => {
+  const v = value ?? {};
+  const patch = (k: keyof Identifier, val: unknown) => onChange({ ...v, [k]: val });
+  return (
+    <div className="grid grid-cols-1 gap-2 rounded border border-slate-200 bg-slate-50 p-2 sm:grid-cols-2">
+      <label>
+        <span className={subLabel}>System</span>
+        <input
+          className={baseField}
+          value={v.system ?? ""}
+          onChange={(e) => patch("system", e.target.value || undefined)}
+        />
+      </label>
+      <label>
+        <span className={subLabel}>Value</span>
+        <input
+          className={baseField}
+          value={v.value ?? ""}
+          onChange={(e) => patch("value", e.target.value || undefined)}
+        />
+      </label>
+    </div>
+  );
+};
